+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 1.1.0
+- Changed display of hints from an inline hint to be in a hover menu.
+- Changed result to include a list of several sample matches.
+
 ## 1.0.0
 - Added inlay hints to preview regular expressions.